requirements.txt
streamlit
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
opencv-python-headless
scikit-fuzzy
Pillow
psutil